+++ test_settings.py
@@ -10,8 +10,11 @@ INSTALLED_APPS = (
     'django.contrib.auth',
     'rest_framework',
     'drf_httpsig',
+    'django_nose',
 )
 
 ROOT_URLCONF = 'drf_httpsig.tests'
 
 SECRET_KEY = 'MY PRIVATE SECRET'
+
+TEST_RUNNER = "django_nose.NoseTestSuiteRunner"