+++ CHANGELOG.rst
@@ -1,15 +1,27 @@
 drf-httpsig Changes
--------------------
+===================
 
-v1.0b2 (2014-Jul-01)
-~~~~~~~~~~~~~~~~~~~~
+v1.0.1 (2014-Jul-03)
+--------------------
+
+* Added/verified Python 3 support and tests (3.2+).
+* Added support for sending a DRF authorization challenge if we're the primary authenticator.
+* Switched to using the `httpsig` HeaderVerifier instead of doing it ourselves. Lots of code got deleted there.
+* Changed fetch_user_data to also receive the algorithm the keyID is for.
+* Updated README.
+* Removed models.py -- the client should handle that part entirely.
+
+v1.0b2/1.0.0 (2014-Jul-01)
+--------------------
+
 * Added versioneer.
 * Updated requirements to use latest httpsig.
 * Added "setup.py test" and tox support.
 * Fixed a unit test.
 
 v1.0b1 (2014-Jun-27)
-~~~~~~~~~~~~~~~~~~~~
+--------------------
+
 * Renamed to drf-httpsig because I don't hate my hands.
 * Updated requirements versions to be more sane.
 * Switched to a different branch for http_signature.
@@ -19,7 +31,7 @@ v1.0b1 (2014-Jun-27)
 
 
 djangorestframework-httpsignature (previous)
---------------------------------------------
+============================================
 
 v0.1.5, 20140613 -- Document installation issue